+++ tools/playerLag/tool.js
@@ -103,6 +103,7 @@
           .addClass('lichessTools-playerLag')
           .appendTo(container);
       }
+      container.siblings('signal').remove();
       const lagRating = !lag ? 0 : lag < 150 ? 4 : lag < 300 ? 3 : lag < 500 ? 2 : 1;
       signal
         .attr('title',trans.noarg('playerLagTitle').replaceAll('%1',Math.round(lag)).replaceAll('%2',Math.round(latency)))